Lots of formatting issues fixed bc res class
The res class basically shrank the size of the paper as
I uploaded to JobMine, so I had to change up a lot of different
formatting. Margins needed to be fixed on the side, and the section
text got modified in the process, which required some invisible
text in order to fix, included in with negative spacing for them
to look nice.